Remove pallBS
